[unitytls] moved conversion functions into separate file
