+++ mcs/class/System/Mono.UnityTLS/Debug.cs
@@ -22,39 +22,6 @@ namespace Mono.Unity
 			throw new TlsException (defaultAlert, message);
 		}
 
-		public static AlertDescription VerifyResultToAlertDescription (UnityTls.unitytls_x509verify_result verifyResult, AlertDescription defaultAlert = AlertDescription.InternalError)
-		{
-			if (verifyResult.HasFlag (UnityTls.unitytls_x509verify_result.UNITYTLS_X509VERIFY_FLAG_EXPIRED))
-				return AlertDescription.CertificateExpired;
-			else if (verifyResult.HasFlag (UnityTls.unitytls_x509verify_result.UNITYTLS_X509VERIFY_FLAG_REVOKED))
-				return AlertDescription.CertificateRevoked;
-			else if (verifyResult.HasFlag (UnityTls.unitytls_x509verify_result.UNITYTLS_X509VERIFY_FLAG_CN_MISMATCH))
-				return AlertDescription.UnknownCA;
-			else if (verifyResult.HasFlag (UnityTls.unitytls_x509verify_result.UNITYTLS_X509VERIFY_FLAG_NOT_TRUSTED))
-				return AlertDescription.CertificateUnknown;
-
-			else if (verifyResult.HasFlag (UnityTls.unitytls_x509verify_result.UNITYTLS_X509VERIFY_FLAG_USER_ERROR1))
-				return AlertDescription.UserCancelled;
-			else if (verifyResult.HasFlag (UnityTls.unitytls_x509verify_result.UNITYTLS_X509VERIFY_FLAG_USER_ERROR2))
-				return AlertDescription.UserCancelled;
-			else if (verifyResult.HasFlag (UnityTls.unitytls_x509verify_result.UNITYTLS_X509VERIFY_FLAG_USER_ERROR2))
-				return AlertDescription.UserCancelled;
-			else if (verifyResult.HasFlag (UnityTls.unitytls_x509verify_result.UNITYTLS_X509VERIFY_FLAG_USER_ERROR3))
-				return AlertDescription.UserCancelled;
-			else if (verifyResult.HasFlag (UnityTls.unitytls_x509verify_result.UNITYTLS_X509VERIFY_FLAG_USER_ERROR4))
-				return AlertDescription.UserCancelled;
-			else if (verifyResult.HasFlag (UnityTls.unitytls_x509verify_result.UNITYTLS_X509VERIFY_FLAG_USER_ERROR5))
-				return AlertDescription.UserCancelled;
-			else if (verifyResult.HasFlag (UnityTls.unitytls_x509verify_result.UNITYTLS_X509VERIFY_FLAG_USER_ERROR6))
-				return AlertDescription.UserCancelled;
-			else if (verifyResult.HasFlag (UnityTls.unitytls_x509verify_result.UNITYTLS_X509VERIFY_FLAG_USER_ERROR7))
-				return AlertDescription.UserCancelled;
-			else if (verifyResult.HasFlag (UnityTls.unitytls_x509verify_result.UNITYTLS_X509VERIFY_FLAG_USER_ERROR8))
-				return AlertDescription.UserCancelled;
-
-			return defaultAlert;
-		}
-
 		public static void CheckAndThrow(UnityTls.unitytls_errorstate errorState, UnityTls.unitytls_x509verify_result verifyResult, string context, AlertDescription defaultAlert = AlertDescription.InternalError)
 		{
 			// Ignore verify result if verification is not the issue.
@@ -63,7 +30,7 @@ namespace Mono.Unity
 				return;
 			}
 
-			AlertDescription alert = VerifyResultToAlertDescription (verifyResult, defaultAlert);
+			AlertDescription alert = UnityTlsConversions.VerifyResultToAlertDescription (verifyResult, defaultAlert);
 			string message = string.Format ("{0} - error code: {1}, verify result: {2}", context, errorState.code, verifyResult);
 			throw new TlsException (alert, message);
 		}
