+++ mcs/class/System/common.sources
@@ -262,6 +262,7 @@ Mono.UnityTls/UnityTls.cs
 Mono.UnityTls/UnityTlsProvider.cs
 Mono.UnityTls/UnityTlsStream.cs
 Mono.UnityTls/UnityTlsContext.cs
+Mono.UnityTls/UnityTlsConversions.cs
 Mono.UnityTls/Debug.cs
 Mono.UnityTls/CertHelper.cs
